updated WebAPI

+++ Aurora/Source/AspireOverflow/Services/UserService.cs
@@ -38,7 +38,7 @@ namespace AspireOverflow.Services
 
         public bool RemoveUser(int UserId)
         {
-            Validation.ValidateId(UserId);
+           if (UserId <= 0) throw new ArgumentException($"User Id must be greater than 0 where UserId:{UserId}");
             try
             {
                 return database.RemoveUser(UserId);
@@ -79,7 +79,7 @@ namespace AspireOverflow.Services
 
         public User GetUsersByID(int UserId)
         {
-            Validation.ValidateId(UserId);
+            if (UserId <= 0) throw new ArgumentException($"User Id must be greater than 0 where UserId:{UserId}");
             try
             {
                 return database.GetUserByID(UserId);
@@ -93,7 +93,7 @@ namespace AspireOverflow.Services
 
         public IEnumerable<User> GetUsersByUserRoleID(int UserRoleID)
         {
-            Validation.ValidateId(UserRoleID);
+            if (UserRoleID <= 0) throw new ArgumentException($"User Role Id must be greater than 0 where UserRoleId:{UserRoleID}");
             try
             {
                 return GetUsersByVerifyStatus(1).Where(user => user.UserRoleId == UserRoleID);
@@ -109,7 +109,8 @@ namespace AspireOverflow.Services
 
         public bool ChangeUserVerificationStatus(int UserId, int VerifyStatusID)
         {
-            Validation.ValidateId(UserId, VerifyStatusID);
+            if (UserId <= 0) throw new ArgumentException($"User Id must be greater than 0  where UserId:{UserId}");
+            if (VerifyStatusID <= 0 && VerifyStatusID > 3) throw new ArgumentException($"Verify Status Id must be between 0 and 3 where VerifyStatusID:{VerifyStatusID}");
             try
             {
                 return database.UpdateUserByVerifyStatus(UserId, VerifyStatusID);
